Relocate SolHelper to Solnet.Programs.Utilities
